+++ src/main/java/com/cxb/storehelperserver/service/StorageService.java
@@ -305,11 +305,8 @@ public class StorageService {
             Calendar calendar = new GregorianCalendar();
             calendar.setTime(storageOrder.getApplyTime());
             calendar.add(Calendar.DATE, -1);
-
-            if (!stockService.delStock(TypeDefine.CommodityType.valueOf(c.getCtype()),
-                    storageOrder.getSid(), c.getCid(), calendar.getTime())) {
-                return RestResult.fail("删除关联商品库存失败");
-            }
+            stockService.delStock(TypeDefine.CommodityType.valueOf(c.getCtype()),
+                    storageOrder.getSid(), c.getCid(), calendar.getTime());
         }
         if (!storageOrderCommodityRepository.delete(oid)) {
             return RestResult.fail("删除关联商品失败");
@@ -318,14 +315,17 @@ public class StorageService {
             return RestResult.fail("删除关联商品附件失败");
         }
 
-        if (!userOrderApplyRepository.delete(TypeDefine.OrderType.STORAGE_IN_ORDER.getValue(), oid)) {
-            return RestResult.fail("删除订单申请人失败");
-        }
-        if (!userOrderReviewRepository.delete(TypeDefine.OrderType.STORAGE_IN_ORDER.getValue(), oid)) {
-            return RestResult.fail("删除订单审核人失败");
-        }
-        if (!userOrderCompleteRepository.delete(TypeDefine.OrderType.STORAGE_IN_ORDER.getValue(), oid)) {
-            return RestResult.fail("删除完成订单失败");
+        if (null == review) {
+            if (!userOrderApplyRepository.delete(TypeDefine.OrderType.STORAGE_IN_ORDER.getValue(), oid)) {
+                return RestResult.fail("删除订单申请人失败");
+            }
+            if (!userOrderReviewRepository.delete(TypeDefine.OrderType.STORAGE_IN_ORDER.getValue(), oid)) {
+                return RestResult.fail("删除订单审核人失败");
+            }
+        } else {
+            if (!userOrderCompleteRepository.delete(TypeDefine.OrderType.STORAGE_IN_ORDER.getValue(), oid)) {
+                return RestResult.fail("删除完成订单失败");
+            }
         }
         if (!storageOrderRepository.delete(oid)) {
             return RestResult.fail("删除订单失败");
@@ -497,11 +497,8 @@ public class StorageService {
             Calendar calendar = new GregorianCalendar();
             calendar.setTime(storageOrder.getApplyTime());
             calendar.add(Calendar.DATE, -1);
-
-            if (!stockService.delStock(TypeDefine.CommodityType.valueOf(c.getCtype()),
-                    storageOrder.getSid(), c.getCid(), calendar.getTime())) {
-                return RestResult.fail("删除关联商品库存失败");
-            }
+            stockService.delStock(TypeDefine.CommodityType.valueOf(c.getCtype()),
+                    storageOrder.getSid(), c.getCid(), calendar.getTime());
         }
         if (!storageOrderCommodityRepository.delete(oid)) {
             return RestResult.fail("删除关联商品失败");
@@ -510,14 +507,17 @@ public class StorageService {
             return RestResult.fail("删除关联商品附件失败");
         }
 
-        if (!userOrderApplyRepository.delete(TypeDefine.OrderType.STORAGE_OUT_ORDER.getValue(), oid)) {
-            return RestResult.fail("删除订单申请人失败");
-        }
-        if (!userOrderReviewRepository.delete(TypeDefine.OrderType.STORAGE_OUT_ORDER.getValue(), oid)) {
-            return RestResult.fail("删除订单审核人失败");
-        }
-        if (!userOrderCompleteRepository.delete(TypeDefine.OrderType.STORAGE_OUT_ORDER.getValue(), oid)) {
-            return RestResult.fail("删除完成订单失败");
+        if (null == review) {
+            if (!userOrderApplyRepository.delete(TypeDefine.OrderType.STORAGE_OUT_ORDER.getValue(), oid)) {
+                return RestResult.fail("删除订单申请人失败");
+            }
+            if (!userOrderReviewRepository.delete(TypeDefine.OrderType.STORAGE_OUT_ORDER.getValue(), oid)) {
+                return RestResult.fail("删除订单审核人失败");
+            }
+        } else {
+            if (!userOrderCompleteRepository.delete(TypeDefine.OrderType.STORAGE_OUT_ORDER.getValue(), oid)) {
+                return RestResult.fail("删除完成订单失败");
+            }
         }
         if (!storageOrderRepository.delete(oid)) {
             return RestResult.fail("删除订单失败");
